Split off state providers for different nodes in separate files
