+++ src/config.ts
@@ -11,9 +11,6 @@ export let config: IConfig = {
   totalLimit: 20 * 1024 * 1024,  // 20 megabytes
   reUploadLimit: 1,
   disableCookie: false,
-  showText: false,
-  showLinks: false,
-  showImages: false,
   sensitiveAttributes: [],
   instrument: false,
   uploadHandler: null,
